fix issue with wind

+++ app/src/main/java/com/raul/androidapps/softwaretesttandem/model/Wind.kt
@@ -4,5 +4,5 @@ package com.raul.androidapps.softwaretesttandem.model
 
 class Wind constructor(
     val speed: Float,
-    val deg: Int
+    val deg: Float
 )
